+++ .gitignore
@@ -1,0 +1,3 @@
+// .gitignore 
+/node_modules
+/dist
