--- src/blocks/place-card/__image/_city/_karachaevsk/place-card__image_city_karachaevsk.css
@@ -1,3 +1,0 @@
-.place-card__image_city_karachaevsk {
-  background-image: url(../../../../../images/card-image.jpg);
-}
