+++ src/index.html
@@ -6,7 +6,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <title>Mesto</title>
-  <link rel="stylesheet" href="./pages/index.css">
+  <link rel="stylesheet" href="<%= htmlWebpackPlugin.files.chunks.main.css %>">
 </head>
 
 <body>
@@ -19,7 +19,7 @@
     <!--preloader.end-->
 
     <header class="header root__section">
-      <img src="./images/logo.svg" alt="mesto logo" class="logo">
+      <img class="logo" src="<%=require('./images/logo.svg')%>" alt="mesto logo">
     </header>
     <div class="profile root__section">
       <div class="user-info">
@@ -111,7 +111,7 @@
   </div>
 
   <!--scripts.begin-->
-  <script src="js/scripts.js?v=1"></script>
+  <script src="<%= htmlWebpackPlugin.files.chunks.main.entry %>"></script>
   <!--scripts.end-->
 
 </body>
